Rename PD hidden helpers to hidden events
