+++ python/sglang/srt/disaggregation/common/__init__.py
@@ -3,4 +3,3 @@ from sglang.srt.disaggregation.common.conn import (
     CommonKVManager,
     CommonKVReceiver,
 )
-from sglang.srt.disaggregation.common.pd_hidden import PDHiddenKVManagerMixin
